Created the first draft for a Trailer
This will most likely take 2-2.5 minutes as a video. The parts in bold inside the curly brackets signify what will be shown during or right after the part before the first bracket. I was not entirely sure for some parts and filled out some of the brackets with somewhat vague information for them.